# Replace parseInt with Number at get 6% boost