--- a/lib/contentTypeParser.js
+++ b/lib/contentTypeParser.js
@@ -189,7 +189,7 @@ function rawBody (request, reply, options, parser, done) {
   const limit = options.limit === null ? parser.bodyLimit : options.limit
   const contentLength = request.headers['content-length'] === undefined
     ? NaN
-    : Number.parseInt(request.headers['content-length'], 10)
+    : Number(request.headers['content-length'])
 
   if (contentLength > limit) {
     reply.send(new FST_ERR_CTP_BODY_TOO_LARGE())