--- a/lib/reply.js
+++ b/lib/reply.js
@@ -292,7 +292,7 @@ Reply.prototype.removeTrailer = function (key) {
 }
 
 Reply.prototype.code = function (code) {
-  const intValue = parseInt(code)
+  const intValue = Number(code)
   if (isNaN(intValue) || intValue < 100 || intValue > 599) {
     throw new FST_ERR_BAD_STATUS_CODE(code || String(code))
   }
@@ -562,7 +562,7 @@ function onSendEnd (reply, payload) {
     const contentLength = reply[kReplyHeaders]['content-length']
     if (!contentLength ||
         (req.raw.method !== 'HEAD' &&
-         parseInt(contentLength, 10) !== Buffer.byteLength(payload)
+         Number(contentLength) !== Buffer.byteLength(payload)
         )
     ) {
       reply[kReplyHeaders]['content-length'] = '' + Buffer.byteLength(payload)