--- a/lib/server.js
+++ b/lib/server.js
@@ -331,7 +331,7 @@ function normalizeListenArgs (args) {
 }
 
 function normalizePort (firstArg) {
-  const port = parseInt(firstArg, 10)
+  const port = Number(firstArg)
   return port >= 0 && !Number.isNaN(port) ? port : 0
 }