@anonrig 2022-09-20
> What version of Node was this tested on?

Latest version of Node 18