@jsumners 2022-09-20
We should have included tests for the inputs we expect. `Number` does not _parse_ the input, it _converts_ the input.