@anonrig 2022-09-20
I'll open a new pull request with the tests. Thanks for the input @jsumners