@Fdawgs 2022-09-21
> I'll open a new pull request with the tests. Thanks for the input @jsumners

Will need to test floats:
```js
Number('1.5') // 1.5
parseInt('1.5', 10) // 1
```