@JoCat 2022-10-17
Offtop
I checked one thing just for fun

```js
var { time, timeEnd } = console
var I = 1e8
var y

time("+num")
for (var i = 0; i < I; i++) y = +i
timeEnd("+num")

time("Number")
for (var i = 0; i < I; i++) y = Number(i)
timeEnd("Number")

time("parseInt")
for (var i = 0; i < I; i++) y = parseInt(i)
timeEnd("parseInt")
```

and got the following result

```
+num: 41.525ms
Number: 56.06ms
parseInt: 226.17ms
```

I use node.js v16.14.0