@JoCat 2022-10-17
For a more accurate comparison, I converted the numbers to strings before processing and got the following result:

```js
var { time, timeEnd } = console
var I = 1e8
var y

time("+num")
for (let i = 0; i < I; i++) y = +`${i}`
timeEnd("+num")

time("Number")
for (let i = 0; i < I; i++) y = Number(`${i}`)
timeEnd("Number")

time("parseInt")
for (let i = 0; i < I; i++) y = parseInt(`${i}`, 10)
timeEnd("parseInt")
```

```
+num: 3.262s
Number: 3.200s
parseInt: 12.480s
```